+++ CHANGELOG.md
@@ -10,12 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - New page for advanced searching, including complex reference filter
+- Translations for entity types
 - Button on edit page to reset or delete rules
 - Warn about unsaved changes when logging out
 - pytest-retry since tests often failing on github (we belief due to low system/cpu performance in CI)
 
 ### Changes
 
+- other rerun lib for pytest
 - updated mex-common to 1.16.0
 - upgrade reflex to 0.8.14
 - updated template to https://github.com/robert-koch-institut/mex-template/commit/81bc5b